backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0 # Para JWT, que usaremos para tokens de sessão
python-multipart==0.0.9         # Para formulários, útil para login
sqlalchemy==2.0.30              # Para interagir com o banco de dados